global: move test to utils/testing

Changelog: None
